move http helper to a file
