+++ server/backend/src/cameras/camera-auth.guard.spec.ts
@@ -1,50 +1,12 @@
 import { CameraAuthGuard } from './camera-auth.guard';
 import { CamerasService } from './cameras.service';
-import { createMock, DeepMocked } from '@golevelup/ts-jest';
-import {
-  ExecutionContext,
-  ForbiddenException,
-  UnauthorizedException,
-} from '@nestjs/common';
+import { createMock } from '@golevelup/ts-jest';
+import { ForbiddenException, UnauthorizedException } from '@nestjs/common';
 import { sign, generateKeyPairSync, verify } from 'crypto';
-import { getMockReq } from '@jest-mock/express';
 import { v4 } from 'uuid';
 import { stringToBytes } from '../util';
-import { Request } from 'express';
 import { NotFoundError } from '../errors.types';
-
-/**
- * Create a mock HTTP execution context that will return a request with the given parameters.
- */
-function createMockHttpExecutionContext(
-  requestOptions?: Parameters<typeof getMockReq>[0],
-): DeepMocked<ExecutionContext> {
-  // '@jest-mock/express' doesn't provide a default header() implementation,
-  // so we need to do that ourselves if we want to pass headers as a dict
-  if (requestOptions && requestOptions.headers && !requestOptions.header) {
-    const mockHeader = jest.fn() as jest.MockedFunction<Request['header']>;
-    mockHeader.mockImplementation((name) => {
-      const header: string | string[] | undefined =
-        requestOptions.headers[name.toLowerCase()];
-      if (header === undefined) {
-        return undefined;
-      } else if (typeof header === 'string') {
-        return header;
-      } else {
-        return header[0];
-      }
-    });
-    requestOptions.header = mockHeader as jest.Mock;
-  }
-  const mockReq = getMockReq(requestOptions);
-  const mockExecutionContext = createMock<ExecutionContext>();
-  mockExecutionContext.switchToHttp.mockReturnValueOnce({
-    getRequest: <T = any>() => mockReq as unknown as T,
-    getNext: () => undefined,
-    getResponse: () => undefined,
-  });
-  return mockExecutionContext;
-}
+import { createMockHttpExecutionContext } from '../../test/helpers/execution-context';
 
 describe(CameraAuthGuard, () => {
   it('should approve requests with valid keys', async () => {
